Move dev views to views folder
